Added translate-function that is more extensive than trans
This function checks if the string exists and if the translated string isn’t empty, if thats true, it also checks if it is the same as the English one. If it is the same as the English one, it logs the translation, so you can look into it if the translation needs an update.

If no translation exists or the translation is empty, it also logs and it returns the string un-translated